"code: set tun mtu and renamed function"
Implemented automatic configuration of the MTU for the TUN interface,
based on the MTU of the bind interface and the `max_addrs` (per MEADcast
packet). This ensures the avoidance of MEADcast packets that exceed the
MTU of the bind interface.

Additionally renamed `get_mdc_pkt_size` to `get_mdc_hdr_size` for
clarity.