code: Improve udp checksum calculation
This commit introduces UDP checksum calculation for MEADcast packets on
the sender. Previously the complete checksum was calculated on the
router performing the MEADcast to unicast transformation. However, above
a certain size the calculated checksum was incorrect. Consequently,
we're now calculating the checksum on the sender and only recalculating
the difference on the router. This fixes the wrong checksum for large
packets and should also improve performance.