code: fix infinite loop during grouping
Previously, when splitting leafs was disabled and the number of leafs of
a router exceeded the maximal number of addresses per packet, an
infinite loop occured.
This commit fixes this bug by splitting leafs into multiple packets if
they exceeded the maximal number of addresses per packet, regardless of
the split flag.